Split solver code into one file per solver class
